tidying geospatial mods
